+++ Api/Api/Controllers/PredictionController.cs
@@ -1,5 +1,11 @@
 using System.Collections.Generic;
+using Api.Dtos;
 using Api.Enums;
+using Api.Exceptions.Base;
+using Api.Exceptions;
+using Api.Services.Predictors;
+using Enum = System.Enum;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using System.Text.Json;
@@ -8,6 +14,12 @@ using Api.Services.Predictors;
 using Microsoft.AspNetCore.Http;
 using Swashbuckle.AspNetCore.Annotations;
 using Enum = System.Enum;
+using Swashbuckle.AspNetCore.Annotations;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using System.Text.Json;
+using System;
 
 namespace Api.Controllers;
 
@@ -34,6 +46,7 @@ public class PredictionController : ControllerBase
     }
     
     [HttpPost]
+    [Route("GRPC")]
     [SwaggerOperation(Summary = "Predicts the outcome of matches, championships, or player performance.", Description = "The input required can differ depending on the type of prediction being made.")]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(int))]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(string))]