+++ Api/Api/Services/Predictors/Implementations/TeamPredictor.cs
@@ -10,22 +10,31 @@ namespace Api.Services.Predictors.Implementations;
 public class TeamPredictor : Predictor
 {
     private readonly ILogger<Predictor> _logger;
-    public TeamPredictor(ILogger<Predictor> logger)
+    private readonly PredictorGrpc.PredictorGrpcClient _predictorClient;
+    public TeamPredictor(ILogger<Predictor> logger, PredictorGrpc.PredictorGrpcClient predictorClient)
     {
         _logger = logger;
+        _predictorClient = predictorClient;
     }
     
     public override int Predict(JsonElement input)
     {
-        TeamInput teamInput;
+        TeamInputDto teamInput;
+        TeamPredictionOutput prediction;
         try
         { 
-            teamInput = JsonConvert.DeserializeObject<TeamInput>(input.GetRawText());
+            teamInput = JsonConvert.DeserializeObject<TeamInputDto>(input.GetRawText());
         }
         catch (Exception e)
         {
             throw new InvalidPredictionInputException();
         }
+
+        //DB Requisition to get the team obj
+        
+        string firstTeamName = "ABCD";
+        string secondTeamName = "EFGH";
+        prediction = _predictorClient.TeamPrediction(new TeamInput { FirstTeamAcronym = firstTeamName, SecondTeamAcronym = secondTeamName });
         
         _logger.LogDebug("Team prediction: {} x {}", teamInput.FirstTeamId, teamInput.SecondTeamId);
         return teamInput.FirstTeamId;