+++ Api/Api/Services/Predictors/PredictorFactory.cs
@@ -14,19 +14,21 @@ public interface IPredictorFactory
 public class PredictorFactory : IPredictorFactory
 {
     private readonly ILogger<Predictor> _logger;
+    private readonly PredictorGrpc.PredictorGrpcClient _predictorClient;
 
-    public PredictorFactory(ILogger<Predictor> logger)
+    public PredictorFactory(ILogger<Predictor> logger, PredictorGrpc.PredictorGrpcClient predictorClient)
     {
         _logger = logger;
+        _predictorClient = predictorClient;
     }
 
     public Predictor Create(PredictionTypes type)
     {
         return type switch
         {
-            PredictionTypes.Championships => new ChampionshipPredictor(_logger),
-            PredictionTypes.Players => new PlayerPredictor(_logger),
-            PredictionTypes.Teams => new TeamPredictor(_logger),
+            PredictionTypes.Championships => new ChampionshipPredictor(_logger, _predictorClient),
+            PredictionTypes.Players => new PlayerPredictor(_logger, _predictorClient),
+            PredictionTypes.Teams => new TeamPredictor(_logger, _predictorClient),
             _ => throw new PredictionTypeNotFoundException()
         };
     }
@@ -36,9 +38,9 @@ public class PredictorFactory : IPredictorFactory
         PredictionTypes enumType = (PredictionTypes) type; 
         return enumType switch
         {
-            PredictionTypes.Championships => new ChampionshipPredictor(_logger),
-            PredictionTypes.Players => new PlayerPredictor(_logger),
-            PredictionTypes.Teams => new TeamPredictor(_logger),
+            PredictionTypes.Championships => new ChampionshipPredictor(_logger, _predictorClient),
+            PredictionTypes.Players => new PlayerPredictor(_logger, _predictorClient),
+            PredictionTypes.Teams => new TeamPredictor(_logger,_predictorClient),
             _ => throw new PredictionTypeNotFoundException(type)
         };
     }